requirements.txt
fastapi
uvicorn
supabase
python-multipart
httpx
groq
python-dotenv
